+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [11.0.0] - 2022-07-05
+## Changes
+- Extend existing `ApiCallException` class with two implementations: `PaymentApiCallException` and `VrpApiCallException`.
+  New error classes may contain additional information about an error such as response code, Open Banking response error 
+  code and a bank error description.
+- Upgrade `com.diffplug.spotless` to `v6.8.0`
+- Upgrade `com.github.spotbugs` to `v5.0.9`
+
 ## [10.0.0] - 2022-05-05
 ## Changes
 - Update `specs/vrp-opeapi.yaml` so that debtorAccount field has OBCashAccountDebtorWithName type